Rename zhipuOpenAI.ts → zhipu.ts
