+++ packages/core/examples/external_clients/zhipu.ts
